moved employee reducers to their own files
