+++ client/src/reducers/employeeReducer.js
@@ -1,0 +1,24 @@
+import {
+  FETCH_EMPLOYEE_FROM_DB_SUCCESS,
+  FETCH_EMPLOYEE_FROM_DB_FAIL
+} from '../actions'
+
+const initialState = {
+  employee: [],
+  error: ''
+}
+
+export const employeeReducer = (state = initialState, action) => {
+  switch (action.type) {
+    case FETCH_EMPLOYEE_FROM_DB_SUCCESS:
+      return {
+        employee: action.payload,
+        error: ''
+      }
+    case FETCH_EMPLOYEE_FROM_DB_FAIL:
+      return { ...initialState, error: 'fetching failed' }
+
+    default:
+      return state
+  }
+}
