+++ client/src/reducers/employeesReducer.js
@@ -1,8 +1,6 @@
 import {
   FETCH_EMPLOYEES_FROM_DB_SUCCESS,
   FETCH_EMPLOYEES_FROM_DB_FAIL,
-  FETCH_EMPLOYEE_FROM_DB_SUCCESS,
-  FETCH_EMPLOYEE_FROM_DB_FAIL,
   UPDATE_TIME_OFF_REQUEST_SUCCESS,
   UPDATE_TIME_OFF_REQUEST_FAIL,
   CREATE_EVENT,
@@ -87,21 +85,6 @@ export const employeesReducer = (state = initialState, action) => {
         })
       }
     }
-
-    default:
-      return state
-  }
-}
-
-export const employeeReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case FETCH_EMPLOYEE_FROM_DB_SUCCESS:
-      return {
-        employee: action.payload,
-        error: ''
-      }
-    case FETCH_EMPLOYEE_FROM_DB_FAIL:
-      return { ...initialState, error: 'fetching failed' }
 
     default:
       return state
